// xdportal/src/app/shared/interface/no-image.interface.ts
// All measure(s) are in px (pixels)
export interface NoImageConfig {
    containerWidth: number;
    containerHeight: number;
    containerBottom: number;
    containerBackground?: string;
    textFontSize: number;
    textLineHeight: number;
    letterSpacing: number;
    parentContainerPadding: number;
}